feat(front): 경로 기반 라우트로 변경
